#31, #18 Made the distance counter more realistic and processed feedback.

+++ src/main/java/nl/tudelft/ti2206/group9/level/State.java
@@ -62,10 +62,9 @@ public final class State {
 	}
 	
 	/**
-	 * 
 	 * @return the distance of the track
 	 */
-	public static int getDistance() {
+	public static double getDistance() {
 		return track.getDistance();
 	}
